refactor: move relay middleware stack to dedicated folder
This unifies the usage of implementable parts moving into a dedicated folder, just like the resource decoders.
